Remove some depreciated helpers, moved and updated DBSUB
